+++ src/modules/events/events.handlers.ts
@@ -9,7 +9,6 @@ import { getLatencyLog } from './events.service';
 import { DsPermission } from '@/types/permissions.types';
 import { permissionsGuard } from '@/modules/guards/guards.service';
 import { getNspEvent, getNspRoomId } from '@/util/helpers';
-import { enqueueMessage } from '@/lib/publisher';
 import { HttpMiddleware, ParsedHttpRequest } from '@/lib/middleware';
 import { BadRequestError } from '@/lib/errors';
 import {
@@ -19,12 +18,13 @@ import {
   verifyTimestamp
 } from '@/modules/auth/auth.service';
 import { getPermissions } from '@/modules/permissions/permissions.service';
+import { addMessageToCache, enqueuePersistenceMessage } from '../history/history.service';
 
 const logger = getLogger('event');
 
 const MAX_TIMESTAMP_DIFF_SECS = 30;
 
-export function handleClientEvent(pgPool: Pool): HttpMiddleware {
+export function handleClientEvent(pgPool: Pool, redisClient: RedisClient): HttpMiddleware {
   return async (res: HttpResponse, req: ParsedHttpRequest) => {
     const requestId = uuid();
 
@@ -94,7 +94,8 @@ export function handleClientEvent(pgPool: Pool): HttpMiddleware {
         message: processedMessageData
       };
 
-      await enqueueMessage(persistedMessageData);
+      await addMessageToCache(logger, redisClient, persistedMessageData);
+      await enqueuePersistenceMessage(logger, persistedMessageData);
 
       res.cork(() =>
         getSuccessResponse(res, {
